--- a/dependencyManagement/build.gradle.kts
+++ b/dependencyManagement/build.gradle.kts
@@ -98,7 +98,6 @@ val DEPENDENCIES = listOf(
   "org.awaitility:awaitility:4.1.0",
   "org.bouncycastle:bcpkix-jdk15on:1.69",
   "org.codehaus.mojo:animal-sniffer-annotations:1.20",
-  "org.curioswitch.curiostack:protobuf-jackson:1.2.0",
   "org.jctools:jctools-core:3.3.0",
   "org.junit-pioneer:junit-pioneer:1.4.2",
   "org.skyscreamer:jsonassert:1.5.0",