--- a/exporters/logging-otlp/build.gradle.kts
+++ b/exporters/logging-otlp/build.gradle.kts
@@ -13,9 +13,8 @@ dependencies {
   compileOnly(project(":sdk:metrics"))
 
   implementation(project(":exporters:otlp:common"))
-  implementation(project(":proto"))
 
-  implementation("org.curioswitch.curiostack:protobuf-jackson")
+  implementation("com.fasterxml.jackson.core:jackson-core")
 
   testImplementation(project(":sdk:testing"))