--- a/exporters/otlp/common/src/jmh/java/io/opentelemetry/exporter/otlp/internal/RequestMarshalBenchmarks.java
+++ b/exporters/otlp/common/src/jmh/java/io/opentelemetry/exporter/otlp/internal/RequestMarshalBenchmarks.java
@@ -5,15 +5,11 @@
 
 package io.opentelemetry.exporter.otlp.internal;
 
-import com.google.protobuf.util.JsonFormat;
 import io.opentelemetry.exporter.otlp.internal.traces.TraceRequestMarshaler;
 import io.opentelemetry.proto.collector.trace.v1.ExportTraceServiceRequest;
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
-import java.io.OutputStreamWriter;
-import java.nio.charset.StandardCharsets;
 import java.util.concurrent.TimeUnit;
-import org.curioswitch.common.protobuf.json.MessageMarshaller;
 import org.openjdk.jmh.annotations.Benchmark;
 import org.openjdk.jmh.annotations.BenchmarkMode;
 import org.openjdk.jmh.annotations.Fork;
@@ -29,12 +25,6 @@ import org.openjdk.jmh.annotations.Warmup;
 @Measurement(iterations = 10, time = 1)
 @Fork(1)
 public class RequestMarshalBenchmarks {
-
-  private static final MessageMarshaller BYTEBUDDY_MARSHALLER =
-      MessageMarshaller.builder()
-          .register(ExportTraceServiceRequest.class)
-          .omittingInsignificantWhitespace(true)
-          .build();
 
   @Benchmark
   @Threads(1)
@@ -81,34 +71,6 @@ public class RequestMarshalBenchmarks {
     TraceRequestMarshaler requestMarshaler = TraceRequestMarshaler.create(state.spanDataList);
     ByteArrayOutputStream customOutput = new ByteArrayOutputStream();
     requestMarshaler.writeJsonTo(customOutput);
-    return customOutput;
-  }
-
-  @Benchmark
-  @Threads(1)
-  public ByteArrayOutputStream marshalJsonProtoByteBuddy(RequestMarshalState state)
-      throws IOException {
-    ExportTraceServiceRequest protoRequest =
-        ExportTraceServiceRequest.newBuilder()
-            .addAllResourceSpans(SpanAdapter.toProtoResourceSpans(state.spanDataList))
-            .build();
-    ByteArrayOutputStream customOutput = new ByteArrayOutputStream();
-    BYTEBUDDY_MARSHALLER.writeValue(protoRequest, customOutput);
-    return customOutput;
-  }
-
-  @Benchmark
-  @Threads(1)
-  public ByteArrayOutputStream marshalJsonProtoReflection(RequestMarshalState state)
-      throws IOException {
-    ExportTraceServiceRequest protoRequest =
-        ExportTraceServiceRequest.newBuilder()
-            .addAllResourceSpans(SpanAdapter.toProtoResourceSpans(state.spanDataList))
-            .build();
-    ByteArrayOutputStream customOutput = new ByteArrayOutputStream();
-    JsonFormat.printer()
-        .omittingInsignificantWhitespace()
-        .appendTo(protoRequest, new OutputStreamWriter(customOutput, StandardCharsets.UTF_8));
     return customOutput;
   }
 }