--- a/exporters/otlp/common/src/main/java/io/opentelemetry/exporter/otlp/internal/Marshaler.java
+++ b/exporters/otlp/common/src/main/java/io/opentelemetry/exporter/otlp/internal/Marshaler.java
@@ -5,6 +5,7 @@
 
 package io.opentelemetry.exporter.otlp.internal;
 
+import com.fasterxml.jackson.core.JsonGenerator;
 import java.io.IOException;
 import java.io.OutputStream;
 
@@ -25,6 +26,13 @@ public abstract class Marshaler {
 
   /** Marshals into the {@link OutputStream} in proto JSON format. */
   public final void writeJsonTo(OutputStream output) throws IOException {
+    try (JsonSerializer serializer = new JsonSerializer(output)) {
+      serializer.writeMessageValue(this);
+    }
+  }
+
+  /** Marshals into the {@link JsonGenerator} in proto JSON format. */
+  public final void writeJsonTo(JsonGenerator output) throws IOException {
     try (JsonSerializer serializer = new JsonSerializer(output)) {
       serializer.writeMessageValue(this);
     }